Api Usage Client refactoring
